feat: WIP ALE linter
This linter is a bit hacky (since it requires multiple executables),
lacks tests, and is hard to advocate for including upstream in ALE,
since the jsonc filetype only exists due to this plugin (so far).

Include it here in the hopes that it may be useful to users of this
plugin.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>